refactor : 음악 목록 갱신 수정
