+++ features/utils/data.ts
@@ -58,55 +58,3 @@ export const tapButton = {
   "tap-silver": "/assets/main/tap-silver.svg",
   "tap-gold": "/assets/main/tap-gold.svg",
 };
-
-export const playListSelection = [
-  {
-    id: 0,
-    name: "크리스마스 캐롤 캐빈",
-    artist: "여경이",
-    youtubeUrlId: "YAsAQgoh52I",
-    thumbnail: "/assets/thumbnail_example.svg",
-    category: "carol",
-  },
-  {
-    id: 1,
-    name: "크리스마스 캐롤2 에센셜",
-    artist: "현지",
-    youtubeUrlId: "a_80o2lDYec",
-    thumbnail: "/assets/thumbnail_example.svg",
-    category: "carol",
-  },
-  {
-    id: 2,
-    name: "디즈니 ost",
-    artist: "픽사",
-    youtubeUrlId: "W_MUcyXz3r4",
-    thumbnail: "/assets/thumbnail_example.svg",
-    category: "love",
-  },
-  {
-    id: 3,
-    name: "지극히 사적인 얘기",
-    artist: "다비치",
-    youtubeUrlId: "7ZcJH33Rksg",
-    thumbnail: "/assets/thumbnail_example.svg",
-    category: "love",
-  },
-  {
-    id: 4,
-    name: "크리스마스 재즈",
-    artist: "혜림이",
-    youtubeUrlId: "NJuSStkIZBg",
-    thumbnail: "/assets/thumbnail_example.svg",
-    category: "carol",
-  },
-
-  {
-    id: 5,
-    name: "k- 캐롤 리스트",
-    artist: "국내 크리스마스 어쩌고",
-    youtubeUrlId: "_j9MnO2tCCo",
-    thumbnail: "/assets/thumbnail_example.svg",
-    category: "carol",
-  },
-];
